Update README.md
High Lights  Watching videos from the intern quest.         
Sharing the fb links to the IT group.       
Sharing the poster adds of teravision in the IT groups where mostly interns can be found.  
Low Lights 
Finding better ways for the task to get more interns or new followers of the page.